refactor: move timestamp_array_to_i64_slice to data mod
